requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dateparser==1.0.0
defusedxml==0.7.1
depocs==1.0.1
diff-match-patch==20200713
Django==2.2
django-admin-interface==0.15.6
django-advanced-filters==1.3.0
django-appconf==1.0.4
django-braces==1.14.0
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-data-browser==4.0.2
django-datatables-view==1.19.1
django-debug-toolbar==3.2
django-debugtools==1.8
django-easy-audit==1.3.0
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
django-js-asset==1.2.2
django-json-widget==1.1.1
django-jsonfield-backport==1.0.3
django-jsonify==0.3.0
django-mptt==0.12.0
django-mptt-admin==2.0.3
django-redis==4.12.1
django-report-builder==6.4.2
django-select2==7.7.1
django-simple-import==2.1.0
django-tabbed-admin==1.0.4
django-tables2==2.3.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
environ==1.0
et-xmlfile==1.0.1
future==0.18.2
hyperlink==21.0.0
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.2
pprintpp==0.4.0
psycopg2-binary==2.8.6
Pygments==2.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
regex==2021.4.4
requests==2.25.1
simplejson==3.17.2
six==1.12.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
termcolor==1.1.0
tzlocal==2.1
urllib3==1.26.4
var-dump==1.2
Werkzeug==1.0.1
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
